% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plots the accidentes within a state}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{The number of state to plot}

\item{year}{The year to plot}
}
\description{
Plots the aciddent locations for a selected state and year
}
\details{
In case an invalid state is selected, the function throws and error. In case there
are no accidents to plot, it shows a warning and no plot is created.
}
\examples{
\dontrun{
fars_map_state(1, 2014)
}
}
