% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Reads a FARS file}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{Name of the FARS file to be read. It may be a plain or a compressed file.}
}
\value{
A tbl_df with the information in the file
}
\description{
Reads data from the US National Highway Traffic Safety Administration's
\href{http://www.nhtsa.gov/Data/Fatality-Analysis-Reporting-System-(FARS)}{Fatality Analysis Reporting System} and stores it into a table.
}
\details{
It returns and error in case the file does not exist
}
\examples{
\dontrun{
# Read data from 2013, in plain csv format
a <- fars_read('accident_2013.csv')

# Read data from 2014, with bz2-compressed csv format
b <- fars_read('accident_2014.csv.bz2')
}
}
