% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read FARS info for a list of years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{List or numeric vector of years to read data from}
}
\value{
A list of tbl_df with the information per year
}
\description{
Reads the FARS info for the list of years passed as parameter, returning them as a list of tables.
}
\details{
In case of error reading the data for one year, it will give a warning and continue with the rest.
}
\examples{

# Reads the data from 2013 to 2015
fars_read_years(2013:2015)
}
