% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Gets the number of accident per month and year}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{List of vector of years to read data from}
}
\value{
A table with one row per month and one column per year. The value of the MONTH column is the month.
The value of the rest of the columns is the number of accident in the year given by the column name.
}
\description{
Reads data FARS data and calculate the number of accidents per month
}
\examples{
\dontrun{
# Reads and summarize data from 2013 to 2015
fars_summarize_years(2013:2015)
}
}
