% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Builds the name of a FARS file}
\usage{
make_filename(year)
}
\arguments{
\item{year}{The year to build the file name}
}
\value{
The name of the FARS file containg the data
}
\description{
Builds the file name of a FARS file. The file name is constructed based on the year
}
\examples{

# Make the filename for year 2001
make_filename(2001)

# Make the filename for year 2016
make_filename(2016)
}
